Move dataflow and eog related queries to own file
